move FieldSetType to v1 namespace
