+++ src/Federation/Extensions/DirectiveTypeDescriptorExtensions.cs
@@ -1,4 +1,5 @@
 using static ApolloGraphQL.HotChocolate.Federation.Constants.WellKnownArgumentNames;
+using FieldSetTypeV1 = ApolloGraphQL.HotChocolate.Federation.One.FieldSetType;
 
 namespace ApolloGraphQL.HotChocolate.Federation;
 
@@ -9,7 +10,7 @@ internal static class DirectiveTypeDescriptorExtensions
     {
         descriptor
             .Argument(Fields)
-            .Type<NonNullType<FieldSetType>>();
+            .Type<NonNullType<FieldSetTypeV1>>();
 
         return descriptor;
     }
